% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{add_metadata}
\alias{add_metadata}
\title{Merge a dataframe containing metadata to a dataframe containing the data.}
\usage{
add_metadata(data, metadata)
}
\arguments{
\item{data}{A dataframe containing data. Should contain a column named "plate_well".}

\item{metadata}{A dataframe containing metadata. Should contain a column named "plate_well".}
}
\value{

}
\description{
Merge a dataframe containing metadata to a dataframe containing the data.
}
