% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{apply_chosen_criteria}
\alias{apply_chosen_criteria}
\title{Apply chosen analyte quality criteria}
\usage{
apply_chosen_criteria(my_data, criteria_to_consider)
}
\arguments{
\item{my_data}{The return value of the function
\code{\link{check_each_criterium}}.}

\item{criteria_to_consider}{Character vector indicating which criteria columns to use.}
}
\value{
The original dataframe my_data with an additional column
  \code{analyte_meets_criteria} that is \code{TRUE} if the analyte fulfills
  all criteria in \code{criteria_to_consider}, otherwise \code{FALSE}.
}
\description{
Determines if the analyte quality criteria in \code{criteria_to_consider} 
were all fulfilled.
}
