% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{calculate_bisection}
\alias{calculate_bisection}
\title{Calculate bisection}
\usage{
calculate_bisection(.data)
}
\arguments{
\item{.data}{A data frame / tibble. Can be passed to the function via a pipe
(\code{\%>\%}) (see Examples below).}
}
\value{
A tibble with the columns \code{sample_name}, \code{cluster},
  \code{group}, \code{Bisection} and \code{bis_formula}.
}
\description{
This function calculates the derived trait bisection based on the relative
abundances of glycans. This function is used in the function
\code{\link{calculate_derived_traits}}. Analytes with "N5" in their name are
considered to be bisected.
}
\examples{
See usage in the function calculate_derived_traits.
}
