% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{calculate_cut_offs}
\alias{calculate_cut_offs}
\title{Calculate cut-off values}
\usage{
calculate_cut_offs(
  summarized_checks,
  control_sample_types = NULL,
  exclude_sample_types = NULL,
  group_keyword = NULL,
  percentile,
  use_mean_SD = FALSE,
  SD_factor = NULL,
  uncalibrated_as_NA = FALSE
)
}
\arguments{
\item{summarized_checks}{The return value of the function
\code{\link{summarize_spectra_checks}}.}

\item{control_sample_types}{The sample type(s) that should be used as
negative control samples to base the cut-offs on. This argument is only
relevant if the method "Based on negative controls" is chosen, otherwise
this argument should be NULL (default).}

\item{exclude_sample_types}{The sample type(s) that should not be taken into
account in the calculation of the cut-off values. This argument is only
relevant when the method "Based on percentiles" is chosen, otherwise this
argument should be NULL (default).}

\item{group_keyword}{If the data contains total and specific Ig samples, the
group_keyword indicates for which group (total or specific) cut-offs are
being calculated. The cut-offs need to be calculated separately because the
control_sample_types can differ between total and specific.}

\item{percentile}{When control_sample_types is not NULL, this is the
percentile of the negative control samples' sum intensity and percentage of
passing analytes that will be used as cut-off values. When
control_sample_types is NULL, this is the percentile of all samples (except
exclude_sample_types)' sum intensity and percentage of passing analytes
that will be used as cut-off values.}

\item{use_mean_SD}{TRUE or FALSE, if TRUE then the mean and standard
deviation (SD) instead of a percentile will be used to calculate the
cut-off value for the sum intensity. The cut-off for the percentage of
passing analytes will still be calculated with a percentile.}

\item{SD_factor}{Only relevant when use_mean_SD is TRUE. The cut-off value
will be calculated as mean + SD_factor * SD.}

\item{uncalibrated_as_NA}{Should uncalibrated spectra be regarded as missing
values (TRUE) or as zeroes (FALSE)?}
}
\value{
A tibble with five or six columns: \describe{\item{group}{This column
  will only exist if the data contained both total and specific samples,
  because in that case the cut-offs are calculated separately for total and
  specific.} \item{cluster}{This column indicates for which cluster the
  cut-off was calculated.} \item{cut_off_sum_intensity}{This numeric column
  gives the cut-off value for the sum intensity.}
  \item{cut_off_passing_analyte_percentage}{This column gives the cut-off
  value for the percentage of passing analytes.} \item{curation_method}{This
  character column can be either "based_on_percentiles" or
  "based_on_negative_controls". It indicates which method for spectra
  curation was used.} \item{sample_type_list}{This column contains a list
  with the sample types that were used as a basis for the analyte curation.}}
}
\description{
\code{calculate_cut_offs()} calculates the cut-off values for the sum
intensity and percentage of passing analytes during spectra curation.
}
