% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{calculate_derived_traits}
\alias{calculate_derived_traits}
\title{Calculate derived glycosylation traits}
\usage{
calculate_derived_traits(data, selected_derived_traits)
}
\arguments{
\item{data}{A dataframe that contains at least the columns "sample_name", "cluster",
"relative_abundance" and "analyte".}

\item{derived_traits}{A character vector containing the names of the derived
traits that should be calculated. The derived traits that can be
calculated with this function are: \describe{ \item{"Fucosylation"}{The
proportion of detected glycans that is fucosylated.} \item{"Bisection"}{The
proportion of detected glycans that contain a bisecting GlcNaC}
\item{"Galactosylation"}{} \item{"Sialylation"}{}}}
}
\value{
A dataframe...
}
\description{
With this function derived glycosylation traits of IgG can be calculated
based on the measured relative abundances of glycans. This function can only
be used on IgG data, because it assumes that all glycans are diantennary.
}
\examples{
example_data <- data.frame(
                   sample_name = c("sample01", "sample02", "sample02", "sample04"),
                   group = c("Spike", "Total", "Spike", "Total"),
                   cluster = c("C1", "C1", "C1", "C2"),
                   analyte = c("H5N4F1", "H4N3F1", "H3N4S1", "H3N4F1S1"),
                   sum_all_analytes = rep(1, 4))
                   
calculate_derived_traits(data = example_data,
                         selected_derived_traits = c("Galactosylation",
                                                     "Bisection"))
}
