% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{calculate_galactosylation}
\alias{calculate_galactosylation}
\title{Calculate galactosylation}
\usage{
calculate_galactosylation(.data)
}
\arguments{
\item{.data}{}
}
\value{

}
\description{
This function calculates the derived trait galactosylation based on the
relative abundances of glycans. This function is used in the function
\code{\link{calculate_derived_traits}}.
}
\examples{
example_data <- data.frame(
                   relative_abundance = c(0.45, 0.15, 0.03, 0.37),
                   analyte = c("H5N4F1", "H4N3F1", "H3N4S1", "H3N4F1S2"),
                   sum_all_analytes = rep(1, 4))
                   
calculate_galactosylation(example_data)

}
