% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_quantitation.R
\name{calculate_IgG1_sum_intensities}
\alias{calculate_IgG1_sum_intensities}
\title{calculate_IgG1_sum_intensities}
\usage{
calculate_IgG1_sum_intensities(
  LaCyTools_summary,
  quantitation_clusters,
  analyte_curated_data
)
}
\arguments{
\item{LaCyTools_summary}{LaCyTools summary, as created in data import.}

\item{quantitation_clusters}{A named list with the cluster names of the peptides that are used
for IgG1 quantitation. This list is created in the data import tab.}

\item{analyte_curated_data}{The analyte curated data from the analyte curation tab.}
}
\value{
A dataframe with the sum intensities of each cluster that is required
for IgG1 quantitation.
}
\description{
calculate_IgG1_sum_intensities
}
