% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_repeatability.R
\name{calculate_repeatability_stats}
\alias{calculate_repeatability_stats}
\title{Calculate the RSD and mean of standards across a plate.}
\usage{
calculate_repeatability_stats(data, standard_sample_id, standard_group)
}
\arguments{
\item{data}{A dataframe with the curated (and normalized) data.}

\item{standard_sample_id}{The sample ID of the replicates that the RSD's
should be calculated for.}

\item{standard_group}{The group (total or specific Ig) that the RSD's should
be calculated for.}
}
\value{
A tibble with 6 columns: \describe{\item{plate}{The plate number.}
  \item{analyte}{The analyte.} \item{average_abundance}{The mean relative
  abundance of that analyte on that plate.} \item{RSD}{The relative standard
  deviation of the relative abundance of that analyte on that plate.}
  \item{n}{The number of samples that the mean and RSD were based on.}
  \item{cluster}{The cluster that the analyte belongs to.}}
}
\description{
Calculate the RSD and mean of standards across a plate.
}
