% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{calculate_sialylation}
\alias{calculate_sialylation}
\title{Calculate sialylation}
\usage{
calculate_sialylation(.data)
}
\arguments{
\item{.data}{A data frame / tibble. Can be passed to the function via a pipe
(\code{\%>\%}) (see Examples below).}
}
\value{
A tibble with the columns \code{sample_name}, \code{cluster},
  \code{group}, \code{Sialylation} and \code{sial_formula}.
}
\description{
This function calculates the derived trait sialylation based on the relative
abundances of glycans. This function is used in the function
\code{\link{calculate_derived_traits}}. It considers analytes with "S1" in
their name to be half sialylated and analytes with "S2" in their name to be
fully sialylated.
}
\examples{
See usage in the function calculate_derived_traits.
}
