% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{calculate_total_intensity}
\alias{calculate_total_intensity}
\title{Calculate the total intensity of an analyte in a spectrum}
\usage{
calculate_total_intensity(data, data_type)
}
\arguments{
\item{data}{The data in long format (one row for each analyte + charge
combination per spectrum). The data should contain at least the columns
\code{absolute_intensity_background_subtracted}, \code{fraction},
\code{sample_name} and \code{analyte}.}

\item{data_type}{"LaCyTools data" or "Skyline data"}
}
\value{
This function returns a tibble with the columns "sample_name",
  "analyte", "cluster", "group" (a factor), "sample_type" (a factor),
  "sample_id", "plate_well", "sum_intensity" (the total intensity of all
  analytes within a spectrum together), "number_of_replicates", "replicates"
  and "total_absolute_intensity" (the total intensity of all charge states of
  an analyte together).
}
\description{
The function \code{calculate_total_intensity} calculates the total intensity
of an analyte in a spectrum, by combining the intensities from each charge
state. For each charge state, the intensity is divided by the fraction. Then
these adjusted intensities are summed up to get the total intensity of the
analyte.
}
