% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{change_axis_title_distance}
\alias{change_axis_title_distance}
\title{Title}
\usage{
change_axis_title_distance(plotly_object, x_distance = 50, y_distance = 50)
}
\arguments{
\item{plotly_object}{}

\item{x_distance}{}

\item{y_distance}{}
}
\description{
Title
}
