% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{check_criteria_skyline}
\alias{check_criteria_skyline}
\title{Check all analyte quality criteria for LaCyTools data.}
\usage{
check_criteria_skyline(
  my_data,
  min_ppm_deviation,
  max_ppm_deviation,
  min_idp,
  min_total_area
)
}
\arguments{
\item{my_data}{A dataframe in long format (one row for each analyte, sample,
and charge combination).}

\item{min_ppm_deviation}{The lowest allowed mass accuracy, in ppm.}

\item{max_ppm_deviation}{The highest allowed mass accuracy, in ppm.}

\item{min_idp}{The lowest allowed isotope dot product value.}

\item{min_total_area}{The lowest allowed total analyte area per charge state.}
}
\value{
The original dataframe, but with an extra column for each quality
  criterium that is \code{TRUE} or \code{FALSE}.
}
\description{
\code{check_criteria_lacytools()} checks for each analyte quality criterium
separately if it is fulfilled, and is used in the functions
\code{\link{check_analyte_quality_criteria_lacytools}} and 
\code{\link{check_analyte_quality_criteria_skyline}}.
}
