% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{check_each_criterium}
\alias{check_each_criterium}
\title{Check all analyte quality criteria}
\usage{
check_each_criterium(
  my_data,
  min_ppm_deviation,
  max_ppm_deviation,
  max_ipq,
  min_sn
)
}
\arguments{
\item{my_data}{A dataframe in long format (one row for each analyte + sample
+ charge combination).}

\item{min_ppm_deviation}{The lowest allowed value for the mass accuracy (in
ppm).}

\item{max_ppm_deviation}{The highest allowed value for the mass accuracy (in
ppm).}

\item{max_ipq}{The highest allowed value for the Isotopic Pattern Quality
(IPQ). The IPQ indicates how much the isotopic pattern deviates from the
theoretic isotopic pattern.}

\item{min_sn}{The lowest allowed value for the signal to noise ratio (S/N).}
}
\value{
The original dataframe, but with an extra column for each quality
  criterium that is \code{TRUE} or \code{FALSE}.
}
\description{
\code{check_each_criterium()} checks for each analyte quality criterium
separately if it is fulfilled, and is used in the function
\code{\link{check_analyte_quality_criteria}}.
}
\examples{
data(example_data)

check_each_criterium(my_data = example_data,
                     min_ppm_deviation = -20,
                     max_ppm_deviation = 20,
                     max_ipq = 0.2,
                     min_sn = 9)

}
