% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{cluster_heatmap}
\alias{cluster_heatmap}
\title{cluster_heatmap}
\usage{
cluster_heatmap(
  normalized_data,
  exclude_sample_types,
  group_facet,
  color_low,
  color_mid,
  color_high,
  color_na
)
}
\arguments{
\item{normalized_data}{Normalized data in long format.}

\item{exclude_sample_types}{Character vector with sample types to exclude. Empty character vector when not applicable.}

\item{group_facet}{Character: name of column that contains biological groups for facets. Use "" when not applicable.}

\item{color_low}{Color of lowest value.}

\item{color_mid}{Color of middle value.}

\item{color_high}{Color of highest value.}

\item{color_na}{Color of background/missing values.}
}
\value{
A ggplot2 heatmap object.
}
\description{
Efficiently creates a heatmap with glycan on x-axis and cluster on y-axis, showing the median
relative abundance for each analyte. Improves performance by pre-filtering data before expensive
separation, sorting, grouping, and summarizing operations. Faceting logic is consolidated and
only performed if necessary. Returns a ggplot2 heatmap or an empty plot if no data is available.
}
