% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{comma_and}
\alias{comma_and}
\title{Collapse strings into one string separated by comma's and "and"}
\usage{
comma_and(string_components)
}
\arguments{
\item{string_components}{A character vector or list with character strings.}
}
\value{
The strings in string_components pasted together, separated by
  comma's. The last two strings are separated by "and".
}
\description{
Collapse strings into one string separated by comma's and "and"
}
\examples{
some_words <- c("apple", "pear", "peach", "raspberry")
comma_and(string_components = some_words)
}
