% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{comma_or}
\alias{comma_or}
\title{Collapse strings into one string separated by comma's and "or"}
\usage{
comma_or(string_components)
}
\arguments{
\item{string_components}{A character vector or list with character strings.}
}
\value{
The strings in string_components pasted together, separated by
  comma's. The last two strings are separated by "or".
}
\description{
Collapse strings into one string separated by comma's and "or"
}
\examples{
some_words <- list("apple", "pear", "peach", "raspberry")
comma_or(string_components = some_words)
}
