% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{create_analyte_curation_table}
\alias{create_analyte_curation_table}
\title{Create a datatable with all analytes that passed curation.}
\usage{
create_analyte_curation_table(dataframe_for_table)
}
\arguments{
\item{dataframe_for_table}{The result of the
\code{\link{prepare_analyte_curation_table}} function.}
}
\value{
A \code{\link[DT]{datatable}} with one row per analyte that passed
  curation, and one column per charge state present in the data to indicate
  whether that charge state of the analyte passed curation.
}
\description{
This function can be used to show the results from the
\code{\link{curate_analytes}} function. It will create a datatable with all
analytes and their charge states that passed curation.
}
