% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{create_curation_data_when_skipped}
\alias{create_curation_data_when_skipped}
\title{Prepare curation data for bar charts when spectra curation is skipped}
\usage{
create_curation_data_when_skipped(summarized_checks)
}
\arguments{
\item{summarized_checks}{The return value of the function
\code{\link{summarize_spectra_checks}}.}
}
\value{
\code{summarized_checks} with two additional columns:
  \describe{\item{has_passed_spectra_curation}{\code{TRUE} for calibrated
  spectra, \code{FALSE} for uncalibrated spectra.}
  \item{reason_for_failure}{\code{NA} for calibrated spectra;
  \code{"Calibration failed."} for uncalibrated spectra.}}
}
\description{
When spectra curation is skipped, this function creates a dataframe
compatible with \code{\link{plot_spectra_curation_results}}. Only
calibration status determines whether a spectrum is considered passing or
failing.
}
