% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{create_expr_ls}
\alias{create_expr_ls}
\title{create_expr_ls}
\usage{
create_expr_ls(str_expr)
}
\description{
Function to transform trait formula string into named list containing the 
expression of the right-hand side of the equation. The name is the left-hand
side of the equation. Used in the "calculate_custrom_trait" function.
Based on:
}
\details{
https://stackoverflow.com/questions/70821721/how-to-use-an-expression-in-dplyrmutate-in-r

expr_rm extracts whatever comes before the "=" sign (spaces around the "=" sign do not matter)
}
