% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{create_formula_list}
\alias{create_formula_list}
\title{create_formula_list}
\usage{
create_formula_list(normalized_data, chosen_traits, chosen_clusters, reference)
}
\arguments{
\item{normalized_data}{normalized_data in long format}

\item{chosen_traits}{Character vector, e.g.  c("fucosylation", "sialylation")}

\item{chosen_clusters}{Character vector, e.g. c("IgGI", "IgGII")}

\item{reference}{Reference file for traits, e.g. human_IgG_ref.}
}
\description{
Creates a list of formulas for human IgG traits.
}
