% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{create_heatmap}
\alias{create_heatmap}
\title{creat_heatmap}
\usage{
create_heatmap(
  normalized_data,
  cluster_name,
  exclude_sample_types,
  color_low,
  color_mid,
  color_high
)
}
\arguments{
\item{normalized_data}{Normalized data in long format.}

\item{cluster_name}{Cluster name (character) for which to make the plot}

\item{exclude_sample_types}{Character vector with sample types to exclude.
Empty vector is not applicable.}

\item{color_low}{Color for lowest value.}

\item{color_mid}{Color for middle value}

\item{color_high}{Color for highest value.}
}
\value{
A ggplot heatmap.
}
\description{
Creates a simple heatmap for normalized data of a cluster.
Sample names are on the y-axis, glycans are on the x-axis.
}
