% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{create_long_data}
\alias{create_long_data}
\title{Transform the Lacytools summary data to a long format.}
\usage{
create_long_data(block, metadata = NULL)
}
\arguments{
\item{block}{A dataframe containing a block from a LacyTools summary file.}

\item{metadata}{A dataframe containing metadata in case the metadata has already
been added to the data.}
}
\value{
A dataframe containing the LacyTools summary in long format.
}
\description{
Transform the Lacytools summary data to a long format.
}
\examples{

}
