% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{curate_analytes_with_list}
\alias{curate_analytes_with_list}
\title{Analyte curation based on a list}
\usage{
curate_analytes_with_list(passing_spectra, analyte_list)
}
\arguments{
\item{passing_spectra}{A dataframe containing the data after spectra curation,
typically the output from \code{\link{remove_unneeded_columns}} or
\code{\link{return_when_spectra_curation_is_skipped}}.}

\item{analyte_list}{The list of analytes that should pass the analyte
curation process. This should be a dataframe or tibble with one column
named "analyte".}
}
\value{
A dataframe with columns \code{cluster}, \code{charge}, \code{analyte},
  and \code{has_passed_analyte_curation}, indicating which analytes are present
  in the provided \code{analyte_list}.
}
\description{
Curate analytes by checking against a predefined list of analytes that should
pass curation. Any analyte present in the list will be marked as passing curation.
}
