% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{curate_spectra}
\alias{curate_spectra}
\title{Perform spectra curation}
\usage{
curate_spectra(checked_data, summarized_checks, cut_offs)
}
\arguments{
\item{checked_data}{The return value of
\code{\link{check_analyte_quality_criteria_lacytools}} or
\code{\link{check_analyte_quality_criteria_skyline}}.}

\item{summarized_checks}{The return value of the function
\code{\link{summarize_spectra_checks}}.}

\item{cut_offs}{The return value of the function
\code{\link{calculate_cut_offs}}.}
}
\value{
The three dataframes given as function arguments joined together,
   with two additional columns: \describe{\item{has_passed_spectra_curation}{\code{TRUE} 
   if the spectrum passed curation, \code{FALSE} otherwise.}
   \item{reason_for_failure}{Describes why the spectrum failed curation, 
   or \code{NA} if it passed.}}
}
\description{
This function performs spectra curation on mass spectrometry data.
All spectra with sum intensities and percentages of passing analytes above
the cut-off values pass spectra curation.
}
