% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{date_with_text}
\alias{date_with_text}
\title{Title}
\usage{
date_with_text(date_text_values, origin = "1899-12-30")
}
\arguments{
\item{date_text_values}{}

\item{origin}{}
}
\value{

}
\description{
Title
}
