% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clusters.R
\name{define_clusters}
\alias{define_clusters}
\title{Define clusters based on analyte names}
\usage{
define_clusters(data, cluster_keywords)
}
\arguments{
\item{data}{A dataframe in long format (one row for each analyte + sample
combination).}

\item{cluster_keywords}{A character vector with keywords
used as regular expressions to classify analytes into clusters.
\code{cluster_keywords} should have the same length as the number of clusters
in the data.}
}
\value{
The function returns the original dataframe given as the data
 argument, but with an additional column named "cluster". This character
 column indicates to what cluster the analyte belongs.
}
\description{
This function determines to which cluster an analyte belongs based on the
analyte name.
}
\examples{
data("example_data")
define_clusters(data = example_data,
                cluster_keywords = "IgGI")
}
