% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{do_criteria_check}
\alias{do_criteria_check}
\title{Perform an analyte quality criteria check for every spectrum in the data.}
\usage{
do_criteria_check(data, min_ppm_deviation, max_ppm_deviation, max_ipq, min_sn)
}
\arguments{
\item{data}{A dataframe in long format (one row for each analyte + sample
combination).}

\item{min_ppm_deviation}{The lowest allowed value for the mass accuracy (in
ppm).}

\item{max_ppm_deviation}{The highest allowed value for the mass accuracy (in
ppm).}

\item{max_ipq}{The highest allowed value for the Isotopic Pattern Quality
(IPQ). The IPQ indicates how much the isotopic pattern deviates from the
theoretic isotopic pattern.}

\item{min_sn}{The lowest allowed value for the signal to noise ratio (S/N).}
}
\value{
A logical vector of which the length corresponds to the number of
  rows in the data. TRUE indicates that the analyte + sample combination
  passed all three quality criteria checks, whereas FALSE indicates one or
  more criteria were not fulfilled.
}
\description{
\code{do_criteria_check()} performs an analyte quality criteria check for
every spectrum in the data. This function is used within the function
\code{\link{summarize_spectra_checks}}.
}
\examples{
data(long_data)
do_criteria_check(data = long_data,
                  min_ppm_deviation = -20,
                  max_ppm_deviation = 20,
                  max_ipq = 0.2,
                  min_sn = 9)
}
