% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example of a processed LacyTools summary}
\format{
A data frame with 6344 rows and 14 variables.
}
\source{
The center for proteomics and metabolomics (CPM) at the Leiden
  university medical center (LUMC)
}
\usage{
example_data
}
\description{
This dataset is an example of a LacyTools summary file that has been read in
and processed using \code{link{read_lacytools_summary}}. This LacyTools
summary contains only measurements of standards and blanks.
}
\keyword{datasets}
