% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_clusters.R
\name{find_cluster_keyword_match}
\alias{find_cluster_keyword_match}
\title{Check if a cluster keyword matches at least one analyte in the data}
\usage{
find_cluster_keyword_match(unique_analytes, cluster_keyword)
}
\arguments{
\item{unique_analytes}{A character vector with all unique analytes in the data.}

\item{cluster_keyword}{A single cluster keyword (a character string).}
}
\value{
A boolean, TRUE if a match was found, FALSE if no match was found.
}
\description{
Check if a cluster keyword matches at least one analyte in the data
}
\examples{
analytes <- c("IgGIH3N4", "IgGIH3N4F1", "IgGIIH5N4F1")

find_cluster_keyword_match(unique_analytes = analytes,
                           cluster_keyword = "IgGII")
                           
}
