% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_lacytools.R
\name{find_widest_row}
\alias{find_widest_row}
\title{Find the widest row in a non-rectangular data file}
\usage{
find_widest_row(path, delim)
}
\arguments{
\item{path}{File path to non-rectangular data.}

\item{delim}{The field separator used in the file.}
}
\value{
The number of fields/columns in the widest line (integer).
}
\description{
Reads a file line-by-line to determine the maximum number of fields in any
single line. This is used to pre-determine the column count needed by
\code{\link{read_non_rectangular}}.
}
