% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{firstlower}
\alias{firstlower}
\title{Convert the first letter of a string to lowercase}
\usage{
firstlower(string)
}
\arguments{
\item{string}{A single character string or a vector with character strings.}
}
\value{
The same string or vector of strings but with the first letter(s) in
  lowercase.
}
\description{
Convert the first letter of a string to lowercase
}
\examples{
firstlower("This sentence starts with LOWERcase.")

some_strings <- c("You can also", 
                  "Use this FUNCTION", 
                  "On a vector of strings")

firstupper(some_strings)

}
