% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_quantitation.R
\name{get_IgG1_quantitation_data}
\alias{get_IgG1_quantitation_data}
\title{get_IgG1_quantitation_data}
\usage{
get_IgG1_quantitation_data(
  LaCyTools_summary,
  quantitation_clusters,
  analyte_curated_data
)
}
\arguments{
\item{LaCyTools_summary}{LaCyTools summary, as created in data import.}

\item{quantitation_clusters}{A named list with the cluster names of the peptides that are used
for IgG1 quantitation. This list is created in the data import tab.}

\item{analyte_curated_data}{The analyte curated data from the analyte curation tab.}
}
\value{
A dataframe with the all non-glycosylated peptide, SILuMAb glycopeptides,
and natural IgG1 glycopeptides. For natural IgG1 glycopeptides, only those
that passed analyte curation are included.
The returned dataframe only contains samples whose IgG1 glycopeptide spectra
passed spectra curation.
}
\description{
get_IgG1_quantitation_data
}
