% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{get_sample_type_options}
\alias{get_sample_type_options}
\title{Get the sample types to put in the menu for negative control samples}
\usage{
get_sample_type_options(summarized_checks, total_or_specific_keyword)
}
\arguments{
\item{summarized_checks}{The return value of the function
\code{\link{summarize_spectra_checks}}.}

\item{total_or_specific_keyword}{A character string with the keyword for
either the total or the specific samples (depending on which menu you are
creating).}
}
\value{
A named character vector. The elements are the sample types and the
  names are the sample types with the \code{total_or_specific_keyword} as
  prefix and "samples" as suffix. The names are what will be shown in the
  selectInput.
}
\description{
Find out which sample types are present in the total or in the specific
samples. These sample types can then be shown in the selection menu for the
negative control samples.
}
