% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{is_truthy}
\alias{is_truthy}
\title{Version of isTruthy that works with non-existing reactive expressions}
\usage{
is_truthy(x)
}
\arguments{
\item{x}{An expression of which you want to test the truthiness.}
}
\value{
If calling the expression x results in an error this function returns
  FALSE. If the expression x can be called this function returns isTruthy(x).
}
\description{
When you use isTruthy on a reactive expression but that reactive doesn't
exist (yet) an error occurs. To avoid that, this function first checks if the
expression x can be called before checking its truthiness.
}
\examples{
# I can't show an example of this outside of a Shiny app.
}
