% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{kick_out_spectra}
\alias{kick_out_spectra}
\title{Filter out spectra that failed spectra curation}
\usage{
kick_out_spectra(curated_spectra)
}
\arguments{
\item{curated_data}{The return value of \code{\link{curate_spectra}}.}
}
\value{
The filtered dataframe given as \code{curated_data}.
}
\description{
This function filters out spectra that failed spectra curation and spectra
that are uncalibrated.
}
