% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{match_human_IgG_traits}
\alias{match_human_IgG_traits}
\title{match_human_IgG_traits}
\usage{
match_human_IgG_traits(human_traits_ui_input)
}
\arguments{
\item{human_traits_ui_input}{Character vector from the UI human IgG traits input.}
}
\value{
A character vector with short names of IgG traits, which correspond to column names
}
\description{
Matches human IgG traits descriptions from UI to column names in human_IgG_ref
}
