% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{match_mouse_IgG_traits}
\alias{match_mouse_IgG_traits}
\title{match_mouse_IgG_traits}
\usage{
match_mouse_IgG_traits(mouse_traits_ui_input)
}
\arguments{
\item{mouse_traits_ui_input}{Character vector from the UI mouse IgG traits input.}
}
\value{
A character vector with short names of mouse IgG traits, which correspond to column names
}
\description{
Matches mouse IgG traits descriptions from UI to column names in mouse_IgG_ref
}
