% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{match_traits}
\alias{match_traits}
\title{match_traits}
\usage{
match_traits(chosen_traits, antibody_type)
}
\arguments{
\item{chosen_traits}{Character vector from the UI traits inputs.}
}
\value{
A character vector with short names of traits, corresponding to column names in reference files.
}
\description{
Matches chosen glycan traits descriptions from UI to column names in in reference files.
}
