% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_exploration.R
\name{my_boxplot}
\alias{my_boxplot}
\title{Create a box plot}
\usage{
my_boxplot(data, xvar, yvar, color = NULL, facets = NULL)
}
\arguments{
\item{data}{The normalized data in a wide format (every analyte has its own
column).}

\item{xvar}{The variable that should be shown on the x-axis.}

\item{yvar}{The variable that should be shown on the y-axis.}

\item{color}{The variable that should correspond to the colors in the plot.
If \code{color} is NULL no variable will be linked to the colors (default).}

\item{facets}{The variable that should be used to facet the plot. If
\code{facets} is NULL the plot will not be faceted (default).}
}
\value{
A ggplot object.
}
\description{
This function creates a box plot with points overlaid on top.
}
