% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{nfacets}
\alias{nfacets}
\title{Determine the number of facets in a ggplot object.}
\usage{
nfacets(ggplot_object)
}
\arguments{
\item{ggplot_object}{A ggplot object.}
}
\value{
An integer that indicates the number of facets in the ggplot.
}
\description{
Determine the number of facets in a ggplot object.
}
\examples{
data(mtcars)
plot <- ggplot2::ggplot(mtcars,
                     ggplot2::aes(x = mpg, y = disp)) +
                     ggplot2::geom_point() +
                     ggplot2::facet_wrap(~ cyl)
                     
nfacets(plot)
}
