% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{normalize_data}
\alias{normalize_data}
\title{Perform total area normalization}
\usage{
normalize_data(total_intensities)
}
\arguments{
\item{total_intensities}{The return value of the
\code{\link{calculate_total_intensity}} function.}
}
\value{
The tibble given as the \code{total_intensities} argument, but with
  the columns \code{total_absolute_intensity} and \code{sum_intensity}
  replaced by the column \code{relative_abundance}.
}
\description{
This function performs total area normalization per cluster for each
spectrum. The total absolute intensity for each analyte per cluster per
spectrum (\code{total_absolute_intensity}) is divided by the sum absolute
intensity of all analytes in that cluster and spectrum
(\code{sum_intensity}). Multiplying this with 100 yields the relative
abundance of the analyte in that cluster and spectrum in percent.
}
