% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{plot_analyte_curation_averages}
\alias{plot_analyte_curation_averages}
\title{Create a heatmap visualization of analyte curation based on average QC parameters}
\usage{
plot_analyte_curation_averages(
  curated_analytes,
  cut_off_averages,
  selected_cluster,
  bio_groups_colname = ""
)
}
\arguments{
\item{curated_analytes}{The result of the \code{\link{curate_analytes}}
function when using average-based curation: A dataframe with all analytes
and charge combinations, their average QC parameter values, and whether or
not they passed analyte curation.}

\item{cut_off_averages}{A list with cut-off values for average quality control
parameters. This is displayed in the tooltip when hovering over tiles.}

\item{selected_cluster}{The cluster for which the analyte curation results
should be plotted.}

\item{bio_groups_colname}{The name of the column that contains the biological groups,
as a character string. If provided (non-empty string), the plot will be faceted by
biological group. Default is an empty string (no grouping).}
}
\value{
A heatmap plot with analytes on the x-axis and charge states on the y-axis.
  Tiles are colored blue for analytes that passed curation and red for those that
  failed. Average values for mass accuracy, IPQ/IDP, and S/N or total area are shown
  in tooltips. If \code{bio_groups_colname} is specified, the plot is faceted by
  biological group.
}
\description{
Create a heatmap showing the results of average-based analyte curation for a single
cluster. Analytes are shown on the x-axis and charge states on the y-axis. Tiles
are colored based on whether the analyte passed or failed curation. Hovering over
tiles displays the average values of all QC parameters.
}
