% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{plot_spectra_curation}
\alias{plot_spectra_curation}
\title{Title}
\usage{
plot_spectra_curation(curated_data, Ig_data)
}
\arguments{
\item{curated_data}{}

\item{Ig_data}{}
}
\value{

}
\description{
Title
}
