% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{plot_spectra_curation_results}
\alias{plot_spectra_curation_results}
\title{Visualize the results of the spectra curation}
\usage{
plot_spectra_curation_results(curated_data, total_and_specific)
}
\arguments{
\item{curated_data}{The return value from the function
\code{\link{curate_spectra}}.}

\item{total_and_specific}{Character string, TRUE or NO. This argument
indicates whether there are total and specific Ig samples in the data.}
}
\value{
A ggplot object.
}
\description{
This function can be used to visualize how many spectra per sample type passed
spectra curation and how many failed and why. It will create a bar plot with 
sample types on the x-axis and the percentage of spectra on the y-axis. Colors 
represent if and why the spectrum failed curation. If the data contains total 
and specific samples, the plot is faceted by group (total or specific). If the 
data contains 4 or less clusters the plot is also faceted by cluster.

The ggplot has a text aesthetic that can be used to show the number of spectra 
and the percentage of spectra as hover info if the ggplot object is converted 
to a ggplotly object (see example).
}
