% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{prepare_analyte_curation_table}
\alias{prepare_analyte_curation_table}
\title{Prepare a dataframe for the analyte curation table}
\usage{
prepare_analyte_curation_table(
  analyte_curated_data,
  selected_cluster,
  by_group
)
}
\arguments{
\item{analyte_curated_data}{Your data after analyte curation, typically
the output from the \code{\link{curate_analytes}} function.}

\item{selected_cluster}{The cluster for which the analyte curation results
should be shown.}

\item{by_group}{Logical indicating whether analyte curation was performed
per biological group. Set to TRUE if analyte curation is done per biological
group, otherwise FALSE.}
}
\value{
This function returns a dataframe that can be passed as the
  \code{dataframe_for_table} argument to the
  \code{\link{create_analyte_curation_table}} function.
}
\description{
Prepare a dataframe for the analyte curation table
}
