% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_sample_ids.R
\name{process_sample_list}
\alias{process_sample_list}
\title{Process a sample list Excel file}
\usage{
process_sample_list(sample_list_file)
}
\arguments{
\item{sample_list_file}{The path to an Excel file containing the sample list.
The Excel file should contain only one sheet with two columns named 
"sample_name" and "sample_id". The sample_name column should contain all
sample names that are present in your LaCyTools summary. The sample_id column
should contain the sample ID corresponding to that sample name. Standards 
and blanks should be included.}
}
\value{
This function returns a dataframe with the contents of the Excel file.
}
\description{
This function reads and processes a sample list Excel file, so that sample 
IDs can be linked to your data.
}
\examples{
path <- system.file("extdata",
                    "Sample_list_example.xlsx",
                    package = "GlycoDash")

process_sample_list(sample_list_file = path)
}
