% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{remove_unneeded_columns}
\alias{remove_unneeded_columns}
\title{Remove columns no longer needed after spectra curation}
\usage{
remove_unneeded_columns(passing_spectra)
}
\arguments{
\item{passing_spectra}{The return value of \code{\link{kick_out_spectra}}.}
}
\value{
The same dataframe given as \code{passing_spectra}, but without the
  columns created during the spectra curation process (except for the
  sum_intensity column).
}
\description{
The \code{sum_intensity} column is the only column from spectra curation that
is kept, because it is needed during normalization.
}
