% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{report_failed_criteria}
\alias{report_failed_criteria}
\title{Report which analyte quality criteria were not met}
\usage{
report_failed_criteria(my_data, criteria_to_consider, data_type)
}
\arguments{
\item{my_data}{The return value of the function
\code{\link{apply_chosen_criteria}}.}

\item{criteria_to_consider}{Character vector of criteria columns to check.}

\item{data_type}{"LaCyTools" or "Skyline", determines which columns to drop.}
}
\value{
The original dataframe my_data with an additional column
  \code{failed_criteria} that describes which criteria from criteria_to_consider
  were not fulfilled. If all criteria were met the value will be \code{NA}.
  Also removes the criteria columns depending on data_type.
}
\description{
Report which analyte quality criteria were not met
}
