% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{return_when_spectra_curation_is_skipped}
\alias{return_when_spectra_curation_is_skipped}
\title{Create the dataframe for next steps when spectra curation is skipped}
\usage{
return_when_spectra_curation_is_skipped(checked_data, summarized_checks)
}
\arguments{
\item{checked_data}{The return value of the function
\code{\link{check_analyte_quality_criteria}}.}

\item{summarized_checks}{The return value of the function
\code{\link{summarize_spectra_checks}}.}
}
\value{
The \code{checked_data} and \code{summarized_checks} dataframes
  joined together, with unneeded columns removed and with uncalibrated
  spectra filtered out.
}
\description{
Create the dataframe for next steps when spectra curation is skipped
}
