% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{sd_p}
\alias{sd_p}
\title{Calculate the population standard deviation.}
\usage{
sd_p(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector or an R object but not a factor coercible to numeric by as.double(x).}

\item{na.rm}{Logical that indicates whether missing values should be removed.}
}
\value{
The population standard deviation (numeric).
}
\description{
Calculate the population standard deviation.
}
\examples{
numbers <- c(2, 5, 3, 9, 6)
sd_p(numbers)

numbers_na <- c(2, 5, 3, 9, 6, NA)
sd_p(numbers, na.rm = TRUE)
}
