% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{summarize_spectra_checks}
\alias{summarize_spectra_checks}
\title{Summarize analyte quality criteria checks}
\usage{
summarize_spectra_checks(checked_data, data_type)
}
\arguments{
\item{checked_data}{The dataframe that is returned by
\code{\link{check_analyte_quality_criteria_lacytools}} or \code{\link{check_analyte_quality_criteria_skyline}}.}

\item{data_type}{"LaCyTools data" or "Skyline data", determined by file input.}
}
\value{
A dataframe that contains one row per spectrum for each cluster.
  The dataframe contains five columns: \describe{
  \item{sample_name}{The name of the sample for which this spectrum was recorded.}
  \item{sample_type}{The type of sample (e.g. negative control, blank etc.).}
  \item{group}{The group (Total or Specific) that this sample belongs to.}
  \item{cluster}{The cluster for which the metrics were calculated.}
  \item{passing_analyte_percentage}{The percentage of analytes that passed the criteria checks in this spectrum.}
  \item{sum_intensity}{The sum intensity of all passing analytes in this spectrum} }
}
\description{
This function calculates the percentage of passing 
analytes per spectrum and the sum intensity of passing analytes per spectrum. 
Should be used after analyte quality criteria checks for every spectrum and 
analyte combination in the data.
}
