% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_isotopes.R
\name{symbol_to_element}
\alias{symbol_to_element}
\title{Convert an element symbol to its full element name}
\usage{
symbol_to_element(symbol)
}
\arguments{
\item{symbol}{A character string with the element symbol (e.g. `"C"`, `"H"`,
`"Na"`).}
}
\value{
A character string with the full element name (e.g. `"carbon"`).
}
\description{
Convert an element symbol to its full element name
}
