% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{throw_out_samples}
\alias{throw_out_samples}
\title{Filter out samples to ignore during analyte curation}
\usage{
throw_out_samples(passing_spectra, samples_to_ignore)
}
\arguments{
\item{passing_spectra}{The return value from the function
\code{\link{remove_unneeded_columns}} or from the function
\code{\link{return_when_spectra_curation_is_skipped}}.}

\item{samples_to_ignore}{A character vector in which every element is either
an group (specific or total) or sample_type that is present in the data
followed by " samples".}
}
\value{
The same dataframe that is given as the \code{passing_spectra}
  argument, but without the rows corresponding to the samples indicated in
  \code{samples_to_ignore}.
}
\description{
With this function samples you can filter out samples that you don't want to
base the analyte curation on.
}
