% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_lacytools.R
\name{transform_skyline_data}
\alias{transform_skyline_data}
\title{transform_skyline_data}
\usage{
transform_skyline_data(raw_skyline_data)
}
\arguments{
\item{raw_skyline_data}{Raw skyline data, read into R with 
the read_skyline_csv() function above.}
}
\value{
A clean dataframe in a similar format as a transformed
LaCyTools summary. The dataframe contains the following columns:
- sample_name
- analyte
- charge
- absolute_intensity_background_subtracted
- mass_accuracy_ppm
- isotope_dot_product
}
\description{
transform_skyline_data
}
