% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_general.R
\name{try_call}
\alias{try_call}
\title{Try to call a reactive expression}
\usage{
try_call(uncalled_reactive_expression)
}
\arguments{
\item{uncalled_reactive_expression}{The reactive expression that you want to
call.}
}
\value{
Either the called reactive expression or NULL if the reactive
  expression was invalidated or did not exist.
}
\description{
This functions tries to call a reactive expression and returns NULL when
there is an error.
}
\examples{
# Can't show an example with reactives outside of a Shiny app.
}
