% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_repeatability.R
\name{visualize_repeatability}
\alias{visualize_repeatability}
\title{Create a plot showing the repeatability per plate}
\usage{
visualize_repeatability(
  repeatability_data,
  selected_group = NULL,
  selected_sample_id
)
}
\arguments{
\item{repeatability_data}{The return value of
\code{\link{calculate_repeatability_stats}}.}

\item{selected_group}{A character string indicating if the repeatability of
the total or of the specific samples should be shown. Default is NULL.}

\item{selected_sample_id}{A character string indicating what sample type
should be shown in the plot.}
}
\value{
A ggplot object.
}
\description{
This function creates a dodged barplot with a bar for each plate, the
relative abundance in percent on the y-axis and the analytes on the x-axis.
The plot is faceted by cluster. On top of each bar a point is shown that
correspond to the relative standard deviation (RSD) in percent.
}
