% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_repeatability.R
\name{visualize_repeatability_mean_bars}
\alias{visualize_repeatability_mean_bars}
\title{Create a plot showing the repeatability for all plates together}
\usage{
visualize_repeatability_mean_bars(
  data,
  selected_sample_id,
  selected_group = NULL
)
}
\arguments{
\item{data}{A dataframe with the curated (and normalized) data.}

\item{selected_sample_id}{A character string indicating what sample type
should be shown in the plot.}

\item{selected_group}{A character string indicating if the repeatability of
the total or of the specific samples should be shown. Default is NULL.}
}
\value{
A ggplot object.
}
\description{
This function creates a bar plot with the relative abundance in percent on
the y-axis and the analytes on the x-axis. On top of the bars points are
shown that correspond to the relative standard deviation (RSD). The plot is
faceted by cluster. Error bars show the standard deviations.
}
