% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{calculate_custom_traits}
\alias{calculate_custom_traits}
\title{Calculate custom derived glycosylation traits}
\usage{
calculate_custom_traits(normalized_data, custom_traits_formulas)
}
\arguments{
\item{normalized_data}{A dataframe that contains at least the columns
"analyte", "plate_well" and "relative_abundance". 
Entries in the column "analyte" must have the form <cluster>1<glycan composition>,
e.g.:

IgGII1H3N4F1.}

\item{custom_traits_formulas}{An Excel file loaded with the "load_excel" function
from the loadxl package. The first column in the Excel file contains the cluster.
The second column contains a trait to calculate for that cluster.
Each custom trait must be places on a new row. 
A formula must have the name of the trait on the left-hand side, 
and an expression on the right-hand side. E.g.:

my_trait = (0.5 * H3N4 + H4N4) / (H3N4F1 + H4N4F1)}
}
\value{
A tibble with the following columns: sample_name, cluster, group, custom traits
and formulas used to calculate custom traits.
}
\description{
Calculate custom derived traits of IgG based on formulas provided in an 
Excel file.
}
