% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{calculate_custom_traits}
\alias{calculate_custom_traits}
\title{calculate_custom_traits}
\usage{
calculate_custom_traits(traits_excel, normalized_data_wide)
}
\arguments{
\item{traits_excel}{An Excel file with two columns: "trait" and "formula".
The "trait" column should contain the names of the traits.
The "formula" column should contain formulas that specify how to calculate the traits.}

\item{normalized_data_wide}{Data frame with normalized data in wide format.}
}
\value{
A wide dataframe with the normalized data + calculated custom traits.
}
\description{
Calculate custom glycosylation traits based on an Excel file
provided by the user.
}
