% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{calculate_cut_offs}
\alias{calculate_cut_offs}
\title{Calculate cut-off values for spectra curation}
\usage{
calculate_cut_offs(spectra_check, cut_off_basis)
}
\arguments{
\item{spectra_check}{The dataframe returned by
\code{\link{summarize_spectra_checks()}}.}

\item{groups_to_filter}{The group (Specific or Total) to base spectra curation
cut-offs on.}

\item{sample_types_to_filter}{The sample type to base spectra curation
cut-offs on.}
}
\value{

}
\description{
Calculate the cutoffs for the proportion of passing analytes per spectrum and
for the sum intensity of passing analytes per spectrum. This function is used
within the function \code{\link{curate_spectra}}.
}
\examples{
data("long_data")
spectra_check <- summarize_spectra_checks(data = long_data,
                                          min_ppm_deviation = -20,
                                          max_ppm_deviation = 20,
                                          max_ipq = 0.2,
                                          min_sn = 9)
calculate_cut_offs(spectra_check = spectra_check,
                   group_to_filter = "Spike",
                   sample_type_to_filter = "CN")
}
