% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{calculate_cut_offs}
\alias{calculate_cut_offs}
\title{Calculate cut-off values for spectra curation}
\usage{
calculate_cut_offs(spectra_check, group_to_filter, sample_type_to_filter)
}
\arguments{
\item{spectra_check}{The dataframe returned by
\code{\link{summarize_spectra_checks()}}.}

\item{group_to_filter}{The group (Specific or Total) to base spectra curation
cut-offs on.}

\item{sample_type_to_filter}{The sample type to base spectra curation
cut-offs on.}
}
\value{

}
\description{
Calculate the cutoffs for the proportion of passing analytes per spectrum and
for the sum intensity of passing analytes per spectrum. This function is used
within the function \code{\link{curate_spectra}}.
}
