% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{calculate_total_intensity}
\alias{calculate_total_intensity}
\title{Calculate the total intensity of an analyte in a spectrum}
\usage{
calculate_total_intensity(data)
}
\arguments{
\item{data}{The data in long format (one row for each analyte + charge
combination per spectrum). The data should contain at least the columns
\code{absolute_intensity_background_subtracted}, \code{fraction},
\code{sample_name} and \code{analyte}.}
}
\value{
This function returns the dataframe given as \code{data} with the
  column "absolute_intensity_background_subtracted" replaced by the column
  "total_absolute_intensity". The dataframe contains one row per analyte per
  sample.
}
\description{
The function \code{calculate_total_intensity} calculates the total intensity
of an analyte in a spectrum, by combining the intensities from each charge
state. For each charge state, the intensity is divided by the fraction.
Then these adjusted intensities are summed up to get the total intensity of
the analyte.
}
\examples{
data("example_data")
calculate_total_intensity(long_data)
}
