% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{calculate_traits}
\alias{calculate_traits}
\title{Automatically calculate traits based on list of formulas}
\usage{
calculate_traits(normalized_data_wide, trait_formulas)
}
\description{
Automatically calculate traits based on list of formulas
}
