% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{check_analyte_quality_criteria_skyline}
\alias{check_analyte_quality_criteria_skyline}
\title{Perform an analyte quality criteria check for every spectrum in Skyline data.}
\usage{
check_analyte_quality_criteria_skyline(
  my_data,
  min_ppm_deviation,
  max_ppm_deviation,
  min_idp,
  min_total_area,
  criteria_to_consider
)
}
\arguments{
\item{my_data}{A dataframe in long format (one row for each analyte + sample
+ charge combination).}

\item{min_ppm_deviation}{The lowest allowed value for the mass accuracy (in
ppm).}

\item{max_ppm_deviation}{The highest allowed value for the mass accuracy (in
ppm).}

\item{min_idp}{The lowest allowed value for the Isotope Dot Product
(IDP). The IDP indicates how much the isotopic pattern deviates from the
theoretic isotopic pattern.}

\item{min_total_area}{The lowest allowed value for the total analyte area (per charge state).}

\item{criteria_to_consider}{A character vector that indicates which quality
criteria should be taken into account during spectra curation.}
}
\value{
The original dataframe given as the my_data argument, but with 3
  additional columns: \describe{\item{analyte_meets_criteria}{A Boolean:
  \code{TRUE} indicates that the analyte + charge combination in that sample fulfilled
  the quality criteria in criteria_to_consider, whereas \code{FALSE} indicates that 
  one or more criteria were not fulfilled.}\item{uncalibrated}{A Boolean:
  \code{TRUE} if the combination of cluster and sample failed calibration in LaCyTools}
  \item{failed_criteria}{A character string describing which criteria in 
  criteria_to_consider were not fulfilled. If all criteria were fulfilled 
  failed_criteria is \code{NA}}}
}
\description{
\code{check_analyte_quality_criteria_skyline()} performs an analyte quality criteria check for
every spectrum in the data.
}
\examples{
data(example_data)

check_analyte_quality_criteria(my_data = example_data,
                               min_ppm_deviation = -20,
                               max_ppm_deviation = 20,
                               min_idp = 0.9,
                               min_total_area = 0,
                               criteria_to_consider = c("Mass accuracy",
                                                        "Isotope dot product",
                                                        "Total area"))
}
