% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_metadata.R
\name{check_column_names}
\alias{check_column_names}
\title{This function checks whether there are forbidden column names in the 
uploaded metadata. Some column names are not allowed because they are
already created in the code, causing conflict.}
\usage{
check_column_names(merged_metadata)
}
\arguments{
\item{merged_metadata}{}
}
\value{
A vector with forbidden column names. Can be length zero.
}
\description{
This function checks whether there are forbidden column names in the 
uploaded metadata. Some column names are not allowed because they are
already created in the code, causing conflict.
}
\examples{
forbidden_colnames <- check_column_names(merged_metadata())
}
