% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{check_criteria_lacytools}
\alias{check_criteria_lacytools}
\title{Check all analyte quality criteria for LaCyTools data.
It is also used for SweetSuite data.}
\usage{
check_criteria_lacytools(
  my_data,
  min_ppm_deviation,
  max_ppm_deviation,
  max_ipq,
  min_sn
)
}
\arguments{
\item{my_data}{A dataframe in long format (one row for each analyte, sample,
and charge combination).}

\item{min_ppm_deviation}{The lowest allowed mass accuracy, in ppm.}

\item{max_ppm_deviation}{The highest allowed mass accuracy, in ppm.}

\item{max_ipq}{The highest allowed isotopic pattern quality value.}

\item{min_sn}{The lowest allowed signal-to-noise ratio.}
}
\value{
The original dataframe, but with an extra column for each quality
  criterium that is \code{TRUE} or \code{FALSE}.
}
\description{
\code{check_criteria_lacytools()} checks for each analyte quality criterium
separately if it is fulfilled, and is used in the functions
\code{\link{check_analyte_quality_criteria_lacytools}} and 
\code{\link{check_analyte_quality_criteria_skyline}}.
}
