% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{check_sample_id_matches}
\alias{check_sample_id_matches}
\title{Title}
\usage{
check_sample_id_matches(plate_design_ids, metadata_ids)
}
\arguments{
\item{plate_design_ids}{}

\item{metadata_ids}{}
}
\value{

}
\description{
Title
}
