% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_lacytools.R
\name{convert_lacytools_summary}
\alias{convert_lacytools_summary}
\title{Convert a LacyTools summary to a 'tidy' dataframe}
\usage{
convert_lacytools_summary(data)
}
\arguments{
\item{data}{A dataframe containing a LacyTools summary returned by
\code{\link{read_non_rectangular}}.}
}
\value{
This function returns a dataframe with the following columns:
 \describe{ \item{sample_name}{The (unchanged) sample names.}
 \item{analyte}{The analyte. For each sample, there is one row per analyte
 per charge state of that analyte.} \item{charge}{The charge state of the
 analyte.} \item{LacyTools output formats}{The dataframe will contain one
 column for each LacyTools output format that was present in the LacyTools
 summary file.} \item{fraction}{The fraction of the isotopic pattern that was
 included for the analyte.} \item{exact_mass}{The exact mass of the most
 abundant isotopologue of the analyte.}}
}
\description{
The function \code{convert_lacytools_summary()} can convert a dataframe 
containing a LacyTools summary to a 'tidy' dataframe in long format. The long 
format means that for each sample there is one row per analyte per charge state.
}
\examples{
data("LacyTools_summary")

convert_lacytools_summary(data = LacyTools_summary)
}
