% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{create_analyte_curation_table}
\alias{create_analyte_curation_table}
\title{Create a datatable with all analytes that passed curation}
\usage{
create_analyte_curation_table(analyte_curated_data, selected_cluster)
}
\arguments{
\item{analyte_curated_data}{Your data filtered so that only the analytes that
passed curation are included.}
}
\value{

}
\description{
Create a datatable with all analytes that passed curation
}
