% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{create_cut_off_plot}
\alias{create_cut_off_plot}
\title{Visualize the spectra curation process}
\usage{
create_cut_off_plot(summarized_checks, color_palette)
}
\arguments{
\item{summarized_checks}{The return value of the function
\code{\link{summarize_spectra_checks}}.}
}
\value{
A ggplot object.
}
\description{
This function can be used to visualize spectra curation. It will create a
scatter plot with the sum intensity plotted against the percentage of passing
analytes. Each point represents one sum spectrum (one cluster from one
sample). Colors represent sample types. Uncalibrated spectra are shown as squares, while
calibrated spectra are shown as dots. If the data contains total and specific
samples, the plot is faceted by group (total or specific).

The ggplot has a text aesthetic that can be used to show as hover
info if the ggplot object is converted to a ggplotly object (see example).
}
