% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{create_cut_off_plot}
\alias{create_cut_off_plot}
\title{Create a plot to show the cut_offs for spectra curation}
\usage{
create_cut_off_plot(summarized_checks)
}
\arguments{
\item{spectra_check}{The dataframe returned by \code{\link{curate_spectra}}.
\code{curate_spectra} returns two dataframes in a list. The dataframe that
should be passed as the \code{spectra_check} argument to
\code{create_cut_off_plot} is named "spectra_check".}
}
\value{
This function returns a ggplot object.
}
\description{
Create a plot to show the cut_offs for spectra curation
}
\examples{
data("example_data")
spectra_curation <- curate_spectra(data = example_data,
                                   clusters_regex = "IgGI1",
                                   min_ppm_deviation = -20,
                                   max_ppm_deviation = 20,
                                   max_ipq = 0.2,
                                   min_sn = 9,
                                   cut_off_basis = c("Spike PBS", "Total PBS"))

create_cut_off_plot(spectra_check = spectra_curation$spectra_check,
                    cut_off_basis = c("Spike PBS", "Total PBS"))
}
