% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{curate_analytes}
\alias{curate_analytes}
\title{Perform analyte curation}
\usage{
curate_analytes(data, cut_off_percentage)
}
\arguments{
\item{data}{The result of the \code{\link{curate_spectra}} function, but
filtered so that only spectra that passed curation remain (see example
below).}

\item{cut_off_percentage}{The minimum percentage of spectra in which an
analyte needs to fulfill the quality criteria in order for that analyte to
pass curation.}

\item{group_to_ignore}{A character string with the keyword for total Ig
samples or the keyword for specific Ig samples. These samples will not be
used as a basis for analyte curation. Specify \code{NULL} to exclude
neither total nor specific Ig samples.}

\item{sample_types_to_ignore}{A character vector with the sample types that
you don't want to base analyte curation on (e.g. standards).}
}
\value{
A dataframe with one row for each combination of analyte and charge
  per cluster, a column named \code{passing_percentage} with the percentage
  of spectra that pass the analyte quality criteria for that analyte and a
  logical column named \code{passed_curation} that is \code{TRUE} for
  analytes with a \code{passing_percentage} above the
  \code{cut_off_percentage} and \code{FALSE} for analytes with a
  \code{passing_percentage} under or equal to the \code{cut_off_percentage}.
}
\description{
Curate analytes based on the percentage of spectra for which the analyte
passes the quality criteria. These quality criteria are the minimum and
maximum allowed mass accuracy deviations, the isotopic pattern quality (IPQ)
and the signal-to-noise ratio (S/N) (see \code{\link{do_criteria_check}}).
Analyte curation should not be based on spectra for which you don't expect
any analytes to pass, or for which you expect all analytes to pass. The
arguments \code{sample_types_to_ignore} and \code{group_to_ignore} should be
used to indicate which spectra should not be used as basis for analyte
curation. This function should be used only after spectra curation has been
performed with \code{\link{curate_spectra}} and after the spectra that did
not pass spectra curation have been filtered out (see example below).
}
\examples{
data("example_data")
example_data <- curate_spectra(data = example_data,
                            min_ppm_deviation = -20,
                            max_ppm_deviation = 20,
                            max_ipq = 0.2,
                            min_sn = 9,
                            clusters_regex = "IgGI1",
                            cut_off_basis = c("Spike PBS", "Total PBS"))

curated_spectra <- example_data$curated_data \%>\%
   dplyr::filter(passed_spectra_curation == TRUE) \%>\% 
   dplyr::select(-passed_spectra_curation)

curate_analytes(data = curated_spectra,
                group_to_ignore = "Total",
                sample_types_to_ignore = c("Visucon", 
                                           "PBS"),
                cut_off_percentage = 25)
                
}
