% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{curate_analytes}
\alias{curate_analytes}
\title{Perform analyte curation}
\usage{
curate_analytes(checked_analytes, cut_off_percentage)
}
\arguments{
\item{checked_analytes}{The result of the
\code{\link{check_analyte_quality_criteria}} function.}

\item{cut_off_percentage}{The minimum percentage of spectra in which an
analyte needs to fulfill the quality criteria in order for that analyte to
pass curation.}
}
\value{
A dataframe with one row for each combination of analyte and charge
  per cluster, a column named \code{passing_percentage} with the percentage
  of spectra that pass the analyte quality criteria for that analyte and a
  logical column named \code{has_passed_analyte_curation} that is \code{TRUE} for
  analytes with a \code{passing_percentage} equal to or above the
  \code{cut_off_percentage} and \code{FALSE} for analytes with a
  \code{passing_percentage} under the \code{cut_off_percentage}.
}
\description{
Curate analytes based on the percentage of spectra in which the analyte
passes the quality criteria. Before this function is used, you should filter
out samples that you don't want to base the analyte curation on with the
function \code{\link{throw_out_samples}} and you should use the function
\code{\link{check_analyte_quality_criteria}} to check the analyte quality
criteria for each analyte in each sample. The reason we check the analyte
quality criteria again even though this already happened during spectra
curation is that during spectra curation one or more criteria could be
disregarded with advanced settings. During analyte curation all three analyte
quality criteria should be considered.
}
\examples{
# First spectra curation has to be performed:
data("example_data")

example_data <- define_clusters(data = example_data,
                                cluster_keywords = "IgGI")

checked_data <- check_analyte_quality_criteria(my_data = example_data,
                                               min_ppm_deviation = -20,
                                               max_ppm_deviation = 20,
                                               max_ipq = 0.2,
                                               min_sn = 9,
                                               criteria_to_consider = c("Mass accuracy",
                                                                        "S/N",
                                                                        "IPQ"))

summarized_checks <- summarize_spectra_checks(checked_data = checked_data)

cut_offs_total <- calculate_cut_offs(summarized_checks = summarized_checks,
                                    control_sample_types = "PBS",
                                    exclude_sample_types = NULL,
                                    group_keyword = "Total",
                                    percentile = 97,
                                    use_mean_SD = FALSE,
                                    SD_factor = NULL,
                                    uncalibrated_as_NA = TRUE)

cut_offs_specific <- calculate_cut_offs(summarized_checks = summarized_checks,
                                        control_sample_types = "PBS",
                                        exclude_sample_types = NULL,
                                        group_keyword = "Spike",
                                        percentile = 97,
                                        use_mean_SD = FALSE,
                                        SD_factor = NULL,
                                        uncalibrated_as_NA = TRUE)

cut_offs <- dplyr::full_join(cut_offs_total,
                             cut_offs_specific)

curated_spectra <- curate_spectra(checked_data = checked_data,
                                  summarized_checks = summarized_checks,
                                  cut_offs = cut_offs)

passing_spectra <- kick_out_spectra(curated_spectra = curated_spectra)

for_analyte_curation <- remove_unneeded_columns(passing_spectra = passing_spectra)

# Then we remove samples that we don't want to base our analyte curation on:
without_samples_to_ignore <- throw_out_samples(
   passing_spectra = for_analyte_curation,
   samples_to_ignore = c("PBS", "Visucon", "IVIGg", "Total")
)

# We check the analyte quality criteria:
checked_analytes <- check_analyte_quality_criteria(my_data = without_samples_to_ignore,
                                                   min_ppm_deviation = -20,
                                                   max_ppm_deviation = 20,
                                                   max_ipq = 0.2,
                                                   min_sn = 9,
                                                   criteria_to_consider = c("Mass accuracy",
                                                                            "S/N",
                                                                            "IPQ"))
                                                                            
curate_analytes(checked_analytes = checked_analytes,
                cut_off_percentage = 25)

}
