% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{curate_analytes_with_list}
\alias{curate_analytes_with_list}
\title{Analyte curation based on a list}
\usage{
curate_analytes_with_list(data, analyte_list)
}
\arguments{
\item{data}{The result of the \code{\link{curate_spectra}} function, but
filtered so that only spectra that passed curation remain (see example
below).}

\item{analyte_list}{The list of analytes that should pass the analyte
curation process. This can be a list with character strings, a character
vector or a dataframe consisting of a single column with character strings.}
}
\value{
This function returns the same dataframe that was given as the
  \code{data} argument, but filtered so that only those analytes that are in
  \code{analyte_list} remain.
}
\description{
Analyte curation based on a list
}
\examples{
data("long_data")
long_data <- curate_spectra(data = long_data,
                            min_ppm_deviation = -20,
                            max_ppm_deviation = 20,
                            max_ipq = 0.2,
                            min_sn = 9,
                            clusters_regex = "IgGI1",
                            group_to_filter = "Spike",
                            sample_type_to_filter = "CN")

curated_spectra <- long_data \%>\%
   dplyr::filter(passed_curation == TRUE)

analyte_list_file <- system.file("inst",
                                 "extdata",
                                 "Analyte_list.xlsx",
                                 package = "glycodash")

analyte_list <- readxl::read_excel(analyte_list_file)

curate_analytes_with_list(data = curated_spectra,
                          analyte_list = analyte_list)
                          
}
