% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{curate_analytes_with_list}
\alias{curate_analytes_with_list}
\title{Analyte curation based on a list}
\usage{
curate_analytes_with_list(data, analyte_list)
}
\arguments{
\item{analyte_list}{The list of analytes that should pass the analyte
curation process. This can be a list with character strings, a character
vector or a dataframe consisting of a single column with character strings.}
}
\value{
This function returns the same dataframe that was given as the
  \code{data} argument, but filtered so that only those analytes that are in
  \code{analyte_list} remain.
}
\description{
Analyte curation based on a list
}
\examples{
data("example_data")
example_data <- curate_spectra(data = example_data,
                            min_ppm_deviation = -20,
                            max_ppm_deviation = 20,
                            max_ipq = 0.2,
                            min_sn = 9,
                            clusters_regex = "IgGI1",
                            cut_off_basis = c("Spike PBS", "Total PBS"))

curated_spectra <- example_data$curated_data \%>\%
   dplyr::filter(has_passed_spectra_curation == TRUE) \%>\% 
   dplyr::select(-has_passed_spectra_curation)

analyte_list_file <- system.file("extdata",
                                 "Analyte_list.xlsx",
                                 package = "glycodash")

analyte_list <- readxl::read_excel(analyte_list_file)

curate_analytes_with_list(data = curated_spectra,
                          analyte_list = analyte_list)
                          
}
