% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{curate_spectra}
\alias{curate_spectra}
\title{Perform spectra curation}
\usage{
curate_spectra(
  data,
  clusters_regex,
  min_ppm_deviation,
  max_ppm_deviation,
  max_ipq,
  min_sn,
  cut_off_basis
)
}
\arguments{
\item{data}{A dataframe in long format (one row for each analyte + sample
combination).}

\item{clusters_regex}{A vector containing character strings. These strings are
used as regular expressions to classify analytes into clusters.
\code{clusters_regex} should contain one character string per cluster in the
data.}

\item{min_ppm_deviation}{The lowest allowed value for the mass accuracy (in
ppm).}

\item{max_ppm_deviation}{The highest allowed value for the mass accuracy (in
ppm).}

\item{max_ipq}{The highest allowed value for the Isotopic Pattern Quality
(IPQ). The IPQ indicates how much the isotopic pattern deviates from the
theoretic isotopic pattern.}

\item{min_sn}{The lowest allowed value for the signal to noise ratio (S/N).}

\item{cut_off_basis}{A character vector or a single character string
specifying which sample types (and optionally which group (Specific or
Total Ig)) the spectra curation cut-offs should be based on. For each
sample type or sample type-group combination that you want to base the
cut-offs on you should add a character string. For example. if you want to
base the cut-offs on all PBS samples and on the Specific negative_control
samples, \code{cut_off_basis} should be \code{c("PBS", "Specific
negative_control")}}
}
\value{
The function returns the original dataframe given as the data
  argument, but with two additional columns. One column is named
  "passed_spectra_curation"; This logical column is \code{TRUE} for spectra that have
  passed curation and \code{FALSE} for spectra that did not pass curation.
  The other column is named "criteria_check" and is \code{TRUE} for analyte +
  sample combinations that passed all three quality criteria checks, whereas
  \code{FALSE} indicates that one or more criteria were not fulfilled.
}
\description{
\code{curate_spectra()} performs spectra curation on mass spectrometry data.
For each spectrum, analytes are curated based on the quality criteria (mass
accuracy, isotopic pattern quality (IPQ) and signal to noise ratio (S/N)).
Then the proportion of passing analytes and the sum intensity of passing
analytes is calculated for each spectrum (using the function
\code{\link{summarize_spectra_checks}}). Based on the average proportion and
sum intensity in a chosen group of samples that should not pass curation
(e.g. Specific Ig negative control samples), cut-off values for spectra
curation are defined (using the function \code{\link{calculate_cut_offs}}).
All spectra with values above those cut-off values pass the spectra curation.
}
\examples{
data("example_data")
curate_spectra(data = example_data,
               clusters_regex = "IgGI1",
               min_ppm_deviation = -20,
               max_ppm_deviation = 20,
               max_ipq = 0.2,
               min_sn = 9,
               cut_off_basis = c("Spike PBS", "Total PBS"))
}
