% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{curate_spectra}
\alias{curate_spectra}
\title{Perform spectra curation}
\usage{
curate_spectra(checked_data, summarized_checks, cut_offs)
}
\value{
The function returns the original dataframe given as the data
  argument, but with two additional columns. One column is named
  "has_passed_spectra_curation"; This logical column is \code{TRUE} for spectra that have
  passed curation and \code{FALSE} for spectra that did not pass curation.
  The other column is named "analyte_meets_criteria" and is \code{TRUE} for analyte +
  sample combinations that passed all three quality criteria checks, whereas
  \code{FALSE} indicates that one or more criteria were not fulfilled.
}
\description{
\code{curate_spectra()} performs spectra curation on mass spectrometry data.
For each spectrum, analytes are curated based on the quality criteria (mass
accuracy, isotopic pattern quality (IPQ) and signal to noise ratio (S/N)).
Then the proportion of passing analytes and the sum intensity of passing
analytes is calculated for each spectrum (using the function
\code{\link{summarize_spectra_checks}}). Based on the average proportion and
sum intensity in a chosen group of samples that should not pass curation
(e.g. Specific Ig negative control samples), cut-off values for spectra
curation are defined (using the function \code{\link{calculate_cut_offs}}).
All spectra with values above those cut-off values pass the spectra curation.
}
\examples{
data("example_data")
curate_spectra(data = example_data,
               clusters_regex = "IgGI1",
               min_ppm_deviation = -20,
               max_ppm_deviation = 20,
               max_ipq = 0.2,
               min_sn = 9,
               cut_off_basis = c("Spike PBS", "Total PBS"))
}
