% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{date_with_text}
\alias{date_with_text}
\title{Read in a metadata file}
\usage{
date_with_text(date_text_values, origin = "1899-12-30")
}
\arguments{
\item{date_text_values}{A column in a dataframe or tibble with serial dates in
the form of character strings.}

\item{origin}{The origin date used to calculate the serial dates. Defaults to
1899-12-30 (year-month-day) which is the origin used in Excel on Windows for
all dates from 1900-01-01 and on.}

\item{metadata_file}{The path to the metadata Excel file.}
}
\value{
This function returns a dataframe with the metadata.


}
\description{
\code{read_metadata()} can be used to read in an Excel file containing
metadata. It reads in all columns as class character and interprets both "NA"
and empty cells as \code{NA}. The values in the first row are converted to
snake case (see: \code{\link[snakecase]{to_snake_case}}) and used as column
names. If the file contains multiple sheets, only the first sheet is read in.
}
\section{Credits}{

This function was based on a solution found on StackOverflow.\cr
Question asked by Ryan C. Thompson:\cr
\url{https://stackoverflow.com/questions/63027147/how-to-read-in-a-date-column-from-an-excel-sheet-with-ambiguous-formats-into-r}\cr
Solution provided by Rui Barradas:\cr
\url{https://stackoverflow.com/users/8245406/rui-barradas}
}

\examples{
path <- system.file("extdata",
                    "Metadata_example.xlsx",
                    package = "glycodash")
                    
read_metadata(path)
Convert serial dates with comments to dates

\code{date_with_text()} tries to convert a vector with serial dates (given as
character strings) to objects of the class "Date". In Excel dates are stored
as serial numbers. These serial dates are the number of days passed since a
set origin date. When Excel data is read into R the serial dates are usually
recognized as dates and converted to objects of class "Date". However, when in
Excel a cell contains both a date and a comment, the column containing this
cell is no longer recognized as class "Date" and is instead converted to class
"Character". \code{date_with_text()} attempts to convert serial dates (given
as character strings) to objects of class "Date". If some serial dates include
comments, the function instead converts the serial dates to character strings
with the dates formatted as "year-month-day". \strong{Important: This function
can only be used within the function \code{\link[dplyr]{across}} from the
dplyr package!}

path <- system.file("extdata",
                    "for_tests",
                    "Dates_with_comment.xlsx",
                    package = "glycodash")

# Dates with comments in Excel are usually interpreted
# as character strings in R:

dates_with_comment <- readxl::read_excel(path)

print(dates_with_comment)

# # A tibble: 6 x 1
# date
# <chr>
# 1 44603
# 2 44604
# 3 13-02-2022 oh no a comment
# 4 44606
# 5 44607
# 6 44608

dates_with_comment <- dates_with_comment \%>\%
   dplyr::mutate(dplyr::across(.cols = date,
                               .fns = date_with_text))

# Because text is found within the dates
# column the following message is shown:
# "Some date entries in date contain text.
# Output will have class character."

print(dates_with_comment)

# # A tibble: 6 x 1
# date
# <chr>
# 1 2022-02-11
# 2 2022-02-12
# 3 13-02-2022 oh no a comment
# 4 2022-02-14
# 5 2022-02-15
# 6 2022-02-16

path <- system.file("extdata",
                    "for_tests",
                    "Dates.xlsx",
                    package = "glycodash")

# Forcing read_excel() to read the dates in as character strings 
# (even though there are no comments in this file) for 
# demonstration purposes:
dates <- readxl::read_excel(path,
                            col_types = "text")

print(dates)

# # A tibble: 6 x 1
# date
# <chr>
# 1 44603
# 2 44604
# 3 44605
# 4 44606
# 5 44607
# 6 44608

dates <- dates \%>\%
   dplyr::mutate(dplyr::across(.cols = date,
                               .fns = date_with_text))
                               
# When no text is found within the dates column,
# the serial dates are converted to objects of class "date":
print(dates)

# # A tibble: 6 x 1
# date
# <date>
# 1 2022-02-11
# 2 2022-02-12
# 3 2022-02-13
# 4 2022-02-14
# 5 2022-02-15
# 6 2022-02-16
          
}
