% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{define_clusters}
\alias{define_clusters}
\title{Define clusters based on analyte names}
\usage{
define_clusters(data, clusters_regex)
}
\arguments{
\item{data}{A dataframe in long format (one row for each analyte + sample
combination).}

\item{clusters_regex}{A vector containing character strings. These strings are
used as regular expressions to classify analytes into clusters.
\code{clusters_regex} should contain one character string per cluster in the
data.}
}
\value{
The function returns the original dataframe given as the data
 argument, but with an additional column named "cluster". This character
 column indicates to what cluster the analyte belongs.
}
\description{
This function determines which cluster an analyte belongs to based on the
analyte name.
}
\examples{
data("example_data")
define_clusters(data = long_data,
                clusters_regex = "IgGI1")
}
