% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{detect_group}
\alias{detect_group}
\title{Detect whether a sample is specific or total Ig based on the sample name.}
\usage{
detect_group(data, keyword_specific, keyword_total)
}
\arguments{
\item{data}{A dataframe of uploaded data containing a column `sample_name`.}

\item{keyword_specific}{The word(s)/characters within the sample name used to
refer to Specific samples.}

\item{keyword_total}{The word(s)/characters within the sample name used to
refer to Total samples.}
}
\value{
The original dataframe plus a new column called `group`, which
indicates whether a sample is specific or total Ig using the specified 
keywords.
}
\description{
Detect whether a sample is specific or total Ig based on the sample name.
}
