% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{detect_plate_and_well}
\alias{detect_plate_and_well}
\title{Detect plate and well}
\usage{
detect_plate_and_well(data)
}
\arguments{
\item{data}{A dataframe. Should include a column named "sample_name".}
}
\value{
The input dataframe with an added column named "plate_well" that says on which plate
and in which well a sample was analysed.
}
\description{
Detect plate and well
}
\examples{
block_example <- data.frame(sample_name = c("s_0216_Specific_pl_3.A10", 
                                            "s_568_Total_P5_H4", 
                                            "pl23_B6.s_8759"), 
                            values = c(13.56, 738.34, 4.56))
detect_plate_and_well(block_example)
}
