% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{find_block}
\alias{find_block}
\title{Find a block in a LacyTools summary file}
\usage{
find_block(data, variable)
}
\arguments{
\item{data}{A dataframe with the LacyTools summary.}

\item{variable}{The name of a LacyTools output format.}
}
\value{
The row indices of the block.
}
\description{
Find a block in a LacyTools summary file
}
