% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_repeatability.R
\name{find_choices_for_repeatability_menu}
\alias{find_choices_for_repeatability_menu}
\title{Determine the choices for the repeatability sample menu}
\usage{
find_choices_for_repeatability_menu(normalized_data)
}
\arguments{
\item{normalized_data}{Description here...}
}
\value{
A list of character strings where each string contains "sample id: "
  followed by the sample ID and if applicable "group: " followed by the total
  or specific keyword. This list can be given as the \code{choices} argument
  to the function \code{\link[shiny]{selectInput}}.
}
\description{
This function finds the sample IDs or if applicable the combinations of
sample ID and group (total or specific) for which there are multiple
measurements in the data.
}
