% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{find_next_na}
\alias{find_next_na}
\title{Find the next empty line from a given line in a LacyTools summary file.}
\usage{
find_next_na(data, row)
}
\arguments{
\item{data}{A dataframe with the LacyTools summary.}

\item{row}{The row used as a starting point from which to search for the next line with NAs.}
}
\value{
The row index for the next line with NAs.
}
\description{
Find the next empty line from a given line in a LacyTools summary file.
}
\examples{
df <- data.frame(c("John", "Lisa", "Paul", NA, "Pete", NA),
                 c(12, 15, 23, NA, 14, NA),
                 c("apple", "pear", "orange", NA, "pear", NA))
find_next_na(data = df,
            row = 2)

find_next_na(data = df,
            row = 5)

}
