% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_lacytools.R
\name{find_widest_row}
\alias{find_widest_row}
\title{Find the widest row in a non-rectangular data file}
\usage{
find_widest_row(path, delim)
}
\arguments{
\item{path}{A path to a file with non-rectangular data.}

\item{delim}{The field separator used in the file. For example, use "\\t" for 
files with tab-separated values and "," or ";" for files with comma-separated 
values (.csv).}
}
\value{
The number of fields/columns in the widest line of the non-rectangular data file (an integer).
}
\description{
\code{find_widest_row()} identifies the row/line in a non-rectangular data file
that contains the highest number of columns/fields. It is used in \code{\link{read_non_rectangular}}
to determine how many columns need to be read in.
}
\examples{
data_file <- system.file("extdata", 
                         "LacyTools_summary_example.txt", 
                         package = "glycodash")
find_widest_row(path = data_file, delim = "\t")
}
