% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{get_analytes_info}
\alias{get_analytes_info}
\title{Get the analytes info from a LacyTools summary for one output format}
\usage{
get_analytes_info(data, variable)
}
\arguments{
\item{data}{A dataframe with a LacyTools summary.}

\item{variable}{The name of a LacyTools output format.}
}
\value{
A dataframe with three columns: analyte, exact mass and fraction. The
  number of rows will correspond to the number of analytes.
}
\description{
This function gets the exact mass of the most abundant isotopologue and the
fraction for each analyte in a LacyTools summary.
}
\examples{
data("LacyTools_summary")
get_analytes_info(data = LacyTools_summary, variable = "S/N (2+)")
}
