% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{get_analytes_info_from_list}
\alias{get_analytes_info_from_list}
\title{Get the analytes info from a LacyTools summary for a list of output formats}
\usage{
get_analytes_info_from_list(data, list_of_variables)
}
\arguments{
\item{data}{A dataframe with a LacyTools summary.}

\item{list_of_variables}{A list/vector with the name of LacyTools output formats.}
}
\value{
A dataframe with three columns (analyte, exact mass and fraction) and one
        row per analyte and charge combination.
}
\description{
Get the analytes info from a LacyTools summary for a list of output formats
}
\examples{
data("LacyTools_summary")
get_analytes_info_from_list(data = LacyTools_summary, list_of_variables = outputs)
}
