% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{get_block}
\alias{get_block}
\title{Get block}
\usage{
get_block(data, variable, name_specific = "Spike", name_total = "Total")
}
\arguments{
\item{data}{}

\item{variable}{}

\item{name_specific}{}

\item{name_total}{}
}
\value{

}
\description{
Get block
}
\examples{
get_block(LacyTools_summary, variable = "Absolute Intensity (Background Subtracted, 2+)")

}
