% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LacyTools_summary}
\alias{LacyTools_summary}
\title{Example of an unprocessed LaCyTools summary}
\format{
A data frame with 1023 rows and 27 columns
}
\source{
The center for proteomics and metabolomics (CPM) at the Leiden
  university medical center (LUMC)
}
\usage{
LacyTools_summary
}
\description{
This dataset is an example of a LaCyTools summary file that has been read in
using \code{link{read_non_rectangular}}, but has not yet been processed
further. This LaCyTools summary contains only measurements of standards and
blanks.
}
\keyword{datasets}
