% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_derived_traits.R
\name{match_traits}
\alias{match_traits}
\title{Matches traits descriptions from UI to column names in traits reference files.}
\usage{
match_traits(traits_ui_input)
}
\arguments{
\item{traits_ui_input}{Character vector from UI traits input}
}
\value{
Character vector with column names of traits that should be calculated,
         matching those in the traits reference files.
}
\description{
Matches traits descriptions from UI to column names in traits reference files.
}
