% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_add_sample_types.R
\name{mod_add_sample_types_ui}
\alias{mod_add_sample_types_ui}
\title{add_sample_types UI Function}
\usage{
mod_add_sample_types_ui(id)
}
\arguments{
\item{id}{The module's namespace identifier.}
}
\description{
A shiny Module to add sample types to your data either 
automatically based on the sample IDs or by uploading a sample types Excel 
file.
}
