% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{plot_analyte_curation}
\alias{plot_analyte_curation}
\title{Create a visual summary of the analyte curation process}
\usage{
plot_analyte_curation(curated_analytes, cut_off_percentage, selected_cluster)
}
\arguments{
\item{curated_analytes}{The result of the \code{\link{curate_analytes}}
function: A dataframe with all analytes and charge combinations in the
data, their percentage of passing spectra and whether or not they passed
analyte curation.}

\item{cut_off_percentage}{The minimum percentage of spectra in which an
analyte needs to fulfill the quality criteria in order for that analyte to
pass curation.}
}
\value{

}
\description{
Create a visual summary of the analyte curation process
}
