% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{plot_spectra_curation_results}
\alias{plot_spectra_curation_results}
\title{Title}
\usage{
plot_spectra_curation_results(curated_data, total_and_specific)
}
\arguments{
\item{curated_data}{}

\item{total_and_specific}{}
}
\description{
Title
}
