% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_sample_ids.R
\name{process_plate_design}
\alias{process_plate_design}
\title{Process the result of read_plate_design()}
\usage{
process_plate_design(plate_design)
}
\arguments{
\item{plate_design}{The dataframe that is returned by the
\code{\link{read_plate_design}} function.}
}
\value{
A dataframe with two columns: \describe{\item{sample_id}{The sample
  IDs as given in the plate design file.} \item{plate_well}{This column
  indicates the plate and well that a sample was analyzed in. The format is
  as follows: the plate number followed by the well ID, separated by an
  underscore (e.g. plate 1 well A1 is 1_A01).}}
}
\description{
This function takes the result from the \code{\link{read_plate_design}}
function and converts it to a different format.
}
