% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{read_analyte_list_file}
\alias{read_analyte_list_file}
\title{Read an analyte list Excel or .rds file}
\usage{
read_analyte_list_file(filepath, filename)
}
\arguments{
\item{filepath}{The path to the Excel or .rds file containing a list of
analytes. If it is an Excel file it should contain only one sheet. If it's
a .rds file it should be a dataframe or tibble. In both cases there should
be one column named "analyte". The analyte column should contain all
analytes that should pass curation.}

\item{filename}{The name of the Excel or .rds file including the file
extension.}
}
\value{
A tibble with one column named "analyte" with the analytes that
  should pass curation.
}
\description{
Read an analyte list Excel or .rds file
}
\examples{
path <- system.file("extdata",
                    "Analyte_list.xlsx",
                    package = "GlycoDash")

read_analyte_list_file(filepath = path,
                      filename = "Analyte_list.xlsx")

}
