% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{read_and_process_plate_design}
\alias{read_and_process_plate_design}
\title{Read and process a plate design file}
\usage{
read_and_process_plate_design(plate_design_file)
}
\arguments{
\item{plate_design_file}{The path to the plate design Excel file. The file
should be in the format described below.}
}
\value{
This function returns a dataframe with four columns:
 \describe{\item{sample_id}{The sample ID's as given in the plate design
 file.} \item{sample_type}{Automatically determined sample types. The
 function takes the first string of letters within the sample ID as the
 sample type. This might not work for your sample ID's so you may need to
 alter the sample types manually after using this function.}
 \item{plate_well}{This column indicated the plate and well that a sample was
 analyzed in. The format is as follows: the plate number followed by the well
 ID, separated by an underscore (e.g. plate 1 well A1 is 1_A01).}
 \item{duplicate}{This column is \code{TRUE} for samples that were specified
 on the plate design as duplicate samples and \code{FALSE} for the remaining
 samples.}}
}
\description{
The function \code{read_and_process_plate_design} reads in a plate design
Excel file and processes it. It automatically determines sample types based
on the sample ID's and finds the samples on a plate design that are specified
as duplicate samples. The sample ID's and sample types of these duplicate
samples are determined by copying the sample ID and type of the preceding
sample.
}
\section{Plate design format}{
 The top-left cell of the Excel sheet should
 contain the plate number (e.g. "Plate 1"). The cells to the right of the
 top-left cell need to be labelled 1-12 (for a 96-well plate), while the
 cells below the top-left cell need to be labelled A-H. The cells within the
 plate should contain the sample ID's.

 \preformatted{ 
 Plate number  1             2            3             ... 
 A            sample_ID_A1  sample_ID_A2  sample_ID_A3 
 B            sample_ID_B1  sample_ID_B2  sample_ID_B3 
 C            sample_ID_C1  sample_ID_C2  sample_ID_C3  
 ...          ...           ...           ...           ...}

 At the bottom of the plate, leave one row blank and then add the next plate
 in the same format.

 Duplicate samples should be indicated in the plate design Excel file as
 "duplicate". \code{read_and_process_plate_design} assumes that duplicate 
 samples are always the duplicate of the preceding sample, that is the 
 sample one well to the left or, in case the duplicate is positioned in the 
 left-most column of the plate, the sample in the right-most well one row up.
}

\examples{
path <- system.file("inst",
                    "extdata",
                    "Plate_design.xlsx",
                    package = "glycodash")

read_and_process_plate_design(plate_design_file = path)
}
