% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_add_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read in one or more metadata files}
\usage{
read_metadata(filepaths, filenames)
}
\arguments{
\item{filepaths}{A list of paths to the metadata files.}

\item{filenames}{A list of filenames (including file extension) corresponding
to the filepaths.}
}
\value{
A named list of dataframes containing the metadata in the files. The
  names correspond to the filenames.
}
\description{
This function can be used to read in a list of metadata files (.xlsx, .xls or
.rds files). In the case of an Excel file empty cells or "NA" cells will be
read in as \code{NA}. All column names will be converted to snake case
(lowercase, words separated by underscores).
}
\examples{
path <- system.file("extdata",
                    "Metadata_example.xlsx",
                    package = "glycodash")

read_metadata(filepaths = list(path),
              filenames = list("Metadata_example.xlsx"))
}
