% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_lacytools.R
\name{read_non_rectangular}
\alias{read_non_rectangular}
\title{Read in non-rectangular delimited files (Optimized)}
\usage{
read_non_rectangular(path, delim = "\\t")
}
\arguments{
\item{path}{A path to a file with non-rectangular data.}

\item{delim}{The field separator used in the file.}
}
\value{
A data.frame containing the data in the flat file.
}
\description{
This optimized version uses data.table::fread for much faster file reading,
especially with large files. It automatically detects the maximum number of
columns and sets appropriate column names.
}
