% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_import.R
\name{read_non_rectangular}
\alias{read_non_rectangular}
\title{Read in non-rectangular delimited files}
\usage{
read_non_rectangular(path, delim = "\\t")
}
\arguments{
\item{path}{A path to a file with non-rectangular data.}

\item{delim}{The field separator used in the file. 
For example, use "\\t" for files with tab-separated values 
and "," or ";" for files with comma-separated values (.csv).}
}
\value{
A data frame (\code{\link[base]{data.frame}}) containing the data in the flat file.
}
\description{
\code{read_non_rectangular()} can read flat files where the number of fields per line is 
not constant (non-rectangular data). Blank lines in the field are not skipped 
and empty fields "" are interpreted as \code{NA}. This function was created to read
LacyTools summary files (tab-delimited .txt files).
}
\examples{
data_file <- system.file("extdata", 
                         "LacyTools_summary_example.txt", 
                         package = "glycodash")
read_non_rectangular(path = data_file, delim = "\t")

}
