% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_lacytools.R
\name{read_skyline_csv}
\alias{read_skyline_csv}
\title{read_skyline_csv}
\usage{
read_skyline_csv(path_to_file)
}
\arguments{
\item{path_to_file}{Path to Skyline CSV file}
}
\value{
Dataframe with raw data read from CSV
}
\description{
read_skyline_csv
}
