% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_data.R
\name{read_sweetsuite_data}
\alias{read_sweetsuite_data}
\title{Read SweetSuite Data}
\usage{
read_sweetsuite_data(datapaths)
}
\arguments{
\item{datapaths}{A character vector of file paths to the SweetSuite output Excel files.}
}
\value{
A dataframe combining the 'Data' tabs from all input Excel files.
}
\description{
This function reads one or more SweetSuite output Excel files.
}
