% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_skyline.R
\name{rename_skyline_isomers}
\alias{rename_skyline_isomers}
\title{Rename isomeric glycan compositions in Skyline data}
\usage{
rename_skyline_isomers(data_renamed_cols)
}
\arguments{
\item{data_renamed_cols}{A dataframe of imported Skyline CSV data (from
\code{\link{read_skyline_csv}}) with columns renamed to \code{"cluster"},
\code{"glycan"} and \code{"charge"}.}
}
\value{
A dataframe with the same structure as \code{data_renamed_cols}, with
  glycan compositions of isomers renamed using \code{"_a"}, \code{"_b"},
  etc.
}
\description{
Detects the presence of isomers in a Skyline CSV file. When an analyte is
present twice in a given charge state, the two duplicate analytes are assumed
to be isomers with the same glycan composition. The glycan compositions are
renamed to distinguish them by appending suffixes \code{"_a"}, \code{"_b"},
etc. A Shiny notification is shown if isomers are detected.
}
