% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_lacytools.R
\name{rename_skyline_isomers}
\alias{rename_skyline_isomers}
\title{rename_skyline_isomers}
\usage{
rename_skyline_isomers(raw_skyline_data, i)
}
\arguments{
\item{raw_skyline_data}{Imported skyline CSV data from the read_skyline_csv() function.}

\item{i}{Number of the Skyline CSV file that is being processed}
}
\value{
Skyline CSV data with glycan compositions of isomers renamed using "_a", "_b", etc.
}
\description{
This function detects the presence of isomers in a Skyline CSV file.
When an analyte is present twice in a given charge state, the two duplicate
analytes are assumed to be isomers with the same glycan composision. The glycan
compositions are renamed to distinguish them.
}
