% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{report_failed_criteria}
\alias{report_failed_criteria}
\title{Report which analyte quality criteria were not met}
\usage{
report_failed_criteria(my_data, criteria_to_consider)
}
\arguments{
\item{my_data}{The return value of the function
\code{\link{apply_chosen_criteria}}.}

\item{criteria_to_consider}{A character vector that indicates which quality
criteria should be taken into account during spectra curation. The analyte 
quality criteria that can be included are "Mass accuracy", "S/N" and "IPQ".}
}
\value{
The original dataframe my_data with an additional column
  \code{failed_criteria} that describes which criteria from criteria_to_consider
  were not fulfilled. If all criteria were met the value will be \code{NA}.
}
\description{
Report which analyte quality criteria were not met
}
\examples{
data(example_data)

checked <- check_each_criterium(my_data = example_data,
                     min_ppm_deviation = -20,
                     max_ppm_deviation = 20,
                     max_ipq = 0.2,
                     min_sn = 9)

applied <- apply_chosen_criteria(my_data = checked,
                                 criteria_to_consider = c("Mass accuracy",
                                                          "S/N",
                                                          "IPQ"))

report_failed_criteria(my_data = applied,
                       criteria_to_consider = c("Mass accuracy",
                                               "S/N",
                                               "IPQ"))
}
