% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_normalization.R
\name{sample_heatmap}
\alias{sample_heatmap}
\title{sample_heatmap}
\usage{
sample_heatmap(
  normalized_data,
  cluster_name,
  exclude_sample_types,
  group_facet,
  color_low,
  color_mid,
  color_high,
  color_na
)
}
\arguments{
\item{normalized_data}{Normalized data in long format.}

\item{cluster_name}{Cluster name (character) for which to make the plot}

\item{exclude_sample_types}{Character vector with sample types to exclude.
Empty vector is not applicable.}

\item{group_facet}{Character: name of column that contains biological groups for facets.
Use an empty character ("") when not applicable.}

\item{color_low}{Color for lowest value.}

\item{color_mid}{Color for middle value}

\item{color_high}{Color for highest value.}

\item{color_na}{Color to for the background / missing values.}
}
\value{
A ggplot heatmap.
}
\description{
Creates a simple heatmap for normalized data. 
Sample names are on the y-axis, glycans are on the x-axis.
Heatmap is created for a specific cluster.
}
