% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{summarize_spectra_checks}
\alias{summarize_spectra_checks}
\title{Summarize analyte quality criteria checks}
\usage{
summarize_spectra_checks(
  data,
  min_ppm_deviation,
  max_ppm_deviation,
  max_ipq,
  min_sn
)
}
\arguments{
\item{data}{A dataframe in long format (one row for each analyte + sample
combination).}

\item{min_ppm_deviation}{The lowest allowed value for the mass accuracy (in
ppm).}

\item{max_ppm_deviation}{The highest allowed value for the mass accuracy (in
ppm).}

\item{max_ipq}{The highest allowed value for the Isotopic Pattern Quality
(IPQ). The IPQ indicates how much the isotopic pattern deviates from the
theoretic isotopic pattern.}

\item{min_sn}{The lowest allowed value for the signal to noise ratio (S/N).}
}
\value{
A dataframe in which each row represents one spectrum. The dataframe
  contains five columns: \describe{ \item{sample_name}{The name of the sample
  for which this spectrum was recorded.} \item{sample_type}{The type of
  sample (e.g. negative control, blank etc.).} \item{group}{The group (Total
  or Specific) that this sample belongs to.} \item{passing_proportion}{The
  proportion of analytes that passed the criteria checks in this spectrum.}
  \item{sum_intensity}{The sum intensity of all passing analytes in this
  spectrum} }
}
\description{
\code{summarize_spectra_checks()} first performs analyte quality criteria
checks for every spectrum in the data using
\code{\link{do_criteria_check}}. Then it calculates the proportion of
passing analytes per spectrum and the sum intensity of passing analytes per
spectrum.
}
\examples{
data("long_data")
summarize_spectra_checks(data = long_data,
                               min_ppm_deviation = -20,
                               max_ppm_deviation = 20,
                               max_ipq = 0.2,
                               min_sn = 9)
}
