% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_spectra_curation.R
\name{summarize_spectra_checks}
\alias{summarize_spectra_checks}
\title{Summarize analyte quality criteria checks}
\usage{
summarize_spectra_checks(data_checked)
}
\value{
A dataframe that contains one row per spectrum for each cluster (
  Thus the number of rows is equal to the number of spectra times the number
  of clusters). The dataframe contains five columns: \describe{
  \item{sample_name}{The name of the sample for which this spectrum was
  recorded.} \item{sample_type}{The type of sample (e.g. negative control,
  blank etc.).} \item{group}{The group (Total or Specific) that this sample
  belongs to.} \item{cluster}{The cluster that the analyte belongs to.}
  \item{passing_proportion}{The proportion of analytes that passed the
  criteria checks in this spectrum.} \item{sum_intensity}{The sum intensity
  of all passing analytes in this spectrum} }
}
\description{
\code{summarize_spectra_checks()} calculates the proportion of passing
analytes per spectrum and the sum intensity of passing analytes per spectrum.
\code{summarize_spectra_checks()} should be used after
\code{\link{do_criteria_check}} has been used to perform analyte quality
criteria checks for every spectrum and analyte combination in the data.
}
\examples{
data("long_data")
summarize_spectra_checks(data = long_data,
                         min_ppm_deviation = -20,
                         max_ppm_deviation = 20,
                         max_ipq = 0.2,
                         min_sn = 9)
}
