% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_analyte_curation.R
\name{throw_out_samples}
\alias{throw_out_samples}
\title{Filter out samples to ignore during analyte curation}
\usage{
throw_out_samples(passing_spectra, samples_to_ignore)
}
\arguments{
\item{passing_spectra}{The return value from the function
\code{\link{curate_spectra}}, but filtered so that the dataframe contains
only the rows for which has_passed_curation == TRUE.}

\item{samples_to_ignore}{A character vector in which element is either a
group (specific or total) or sample_type that is present in the data
followed by " samples".}
}
\description{
With this function samples you can filter out samples that you don't want to
base the analyte curation on.
}
