% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_lacytools.R
\name{transform_skyline_data_wide}
\alias{transform_skyline_data_wide}
\title{transform_skyline_data_wide}
\usage{
transform_skyline_data_wide(
  raw_skyline_data_wide,
  analyte_colname = NULL,
  cluster_colname = NULL,
  glycan_colname = NULL,
  charge_colname
)
}
\arguments{
\item{raw_skyline_data_wide}{Raw skyline data in wide format, read into R with 
the read_skyline_csv() function above.}
}
\value{
A clean dataframe in a similar format as a transformed
LaCyTools summary. The dataframe contains the following columns:
- sample_name
- analyte
- charge
- absolute_intensity_background_subtracted
- mass_accuracy_ppm
- isotope_dot_product
}
\description{
transform_skyline_data_wide
}
