% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_repeatability.R
\name{visualize_repeatability}
\alias{visualize_repeatability}
\title{Create a plot that visualizes the repeatability}
\usage{
visualize_repeatability(repeatability_data)
}
\arguments{
\item{repeatability_data}{The dataframe that is returned by the
\code{\link{calculate_repeatability_stats}} function.}
}
\description{
Create a plot that visualizes the repeatability
}
